# management _sysytem/app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App Name API Documentation",
...
        "version": "1.0.0",
        "x-logo": {
            "url": "http://laravel-project-template.test/logo.png",
            "backgroundColor": "#FFFFFF",
            "altText": "LDE"
        }
    },
    "servers": [
        {
            "url": "http://laravel-project-template.test",
            "description": "Production Server"
        },
        {
            "url": "http://laravel-project-template.test",
            "description": "Staging Server"
        },
        {
            "url": "http://laravel-project-template.test/api/v1"
        }
    ],
    "paths": {
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "Logs out current user",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Sucessfull Logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Successfully logged out"
                                        },
                                        "data": {
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response4xx"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/me": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get name of logged in user",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Successfull Retrieval",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Username retreived successfully"
                                        },
                                        "data": {
                                            "example": {
                                                "user_name": "Asia Walter"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response4xx"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/groups": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "This API returns groups in which user has",
                "operationId": "group",
                "responses": {
                    "200": {
                        "description": "Group retrived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Shows operation status",
                                            "type": "boolean",
                                            "default": true
                                        },
                                        "message": {
                                            "description": "Shows operation message",
                                            "type": "string",
                                            "default": "Group retrived successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "groups": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response5xx"
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Client side error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response4xx"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "response4xx": {
                "description": "Response for 4xx type of errors",
                "properties": {
                    "success": {
                        "description": "Shows status of operation",
                        "type": "boolean",
                        "default": false
                    },
                    "message": {
                        "description": "Error message to be shown to user",
                        "type": "string",
                        "example": "Record not found."
                    },
                    "error": {
                        "description": "Full error for debugging",
                        "example": {
                            "key1": "data1",
                            "key2": "data2"
                        }
                    }
                },
                "type": "object"
            },
            "response5xx": {
                "description": "Response for 5xx type of errors",
                "properties": {
                    "success": {
                        "description": "Shows status of operation",
                        "type": "boolean",
                        "default": false
                    },
                    "message": {
                        "description": "Error message to be shown to user",
                        "type": "string",
                        "example": "Database error."
                    },
                    "error": {
                        "description": "Full error for debugging",
                        "example": {
                            "key1": "data1",
                            "key2": "data2"
                        }
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "Header": {
                "type": "apiKey",
                "description": "This API uses TOKEN KEY authentication\n\nKey MUST be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "header",
                "scheme": "https"
            },
            "Token": {
                "type": "apiKey",
                "description": "This API uses **TOKEN KEY** authentication\n\nKey `MUST` be provided for each request that requires authentication.",
                "name": "apikey",
                "in": "query",
                "scheme": "https"
            }
        }
    },
...
}